requirements.txt
asgiref==3.3.4
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.4
django-allauth==0.44.0
django-crispy-forms==1.12.0
django-on-heroku==1.1.2
django-ranged-response==0.2.0
django-simple-captcha==0.5.14
gunicorn==20.1.0
idna==2.10
isodate==0.6.0
oauthlib==3.1.1
Pillow==8.2.0
psycopg2-binary==2.9.1
pycparser==2.20
PyJWT==2.1.0
PySocks==1.7.1
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
selenium==3.141.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
tweepy==3.10.0
urllib3==1.26.5
whitenoise==5.2.0
